% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_corr.r
\name{sample_corr}
\alias{sample_corr}
\title{Draw from the sampling distribution of the correlation coefficient}
\usage{
sample_corr(hat_r, sample_size, n_draws)
}
\description{
Draw from the sampling distribution of the correlation coefficient
}
